+++ tla-bmcmt/src/test/scala/at/forsyte/apalache/tla/bmcmt/TestSymbStateRewriterSet.scala
@@ -859,86 +859,6 @@ class TestSymbStateRewriterSet extends RewriterBase with TestingPredefs {
     }
   }
 
-  test("""SE-SET-CAP[1-2]: {1, 3} \cap {3, 4} = {3}""") {
-    def mkSet(elems: TlaEx*) = OperEx(TlaSetOper.enumSet, elems: _*)
-
-    val left = mkSet(ValEx(TlaInt(1)), ValEx(TlaInt(3)))
-    val right = mkSet(ValEx(TlaInt(3)), ValEx(TlaInt(4)))
-    val expected = mkSet(ValEx(TlaInt(3)))
-    val capSet = OperEx(TlaSetOper.cap, left, right)
-    val eqExpected = OperEx(TlaOper.eq, capSet, expected)
-
-    val state = new SymbState(eqExpected, BoolTheory(), arena, new Binding)
-    val rewriter = create()
-    val nextState = rewriter.rewriteUntilDone(state)
-    nextState.ex match {
-      case predEx@NameEx(name) =>
-        assert(BoolTheory().hasConst(name))
-        assert(solverContext.sat())
-        // check equality
-        rewriter.push()
-        solverContext.assertGroundExpr(predEx)
-        assert(solverContext.sat())
-        rewriter.pop()
-        solverContext.assertGroundExpr(OperEx(TlaBoolOper.not, predEx))
-        assert(!solverContext.sat())
-
-      case _ =>
-        fail("Unexpected rewriting result")
-    }
-  }
-
-  test("""SE-SET-DIFF[1-2]: {1, 3, 5} \cap {1, 4} = {3, 5}""") {
-    def mkSet(elems: TlaEx*) = OperEx(TlaSetOper.enumSet, elems: _*)
-
-    val left = mkSet(ValEx(TlaInt(1)), ValEx(TlaInt(3)), ValEx(TlaInt(5)))
-    val right = mkSet(ValEx(TlaInt(1)), ValEx(TlaInt(4)))
-    val expected = mkSet(ValEx(TlaInt(3)), ValEx(TlaInt(5)))
-    val minusSet = OperEx(TlaSetOper.setminus, left, right)
-    val eqExpected = OperEx(TlaOper.eq, minusSet, expected)
-
-    val state = new SymbState(eqExpected, BoolTheory(), arena, new Binding)
-    val rewriter = create()
-    val nextState = rewriter.rewriteUntilDone(state)
-    nextState.ex match {
-      case predEx@NameEx(name) =>
-        assert(BoolTheory().hasConst(name))
-        assert(solverContext.sat())
-        // check equality
-        rewriter.push()
-        solverContext.assertGroundExpr(predEx)
-        assert(solverContext.sat())
-        rewriter.pop()
-        solverContext.assertGroundExpr(OperEx(TlaBoolOper.not, predEx))
-        assert(!solverContext.sat())
-
-      case _ =>
-        fail("Unexpected rewriting result")
-    }
-  }
-
-  test("""SE-SET-CUP: regression""") {
-    // 2019-01-18, Igor: this bug originally appeared in TwoPhase.tla, the MWE can be found in Bug20190118.tla
-    // S = {1} \ {1}
-    val set1 = tla.setminus(tla.enumSet(tla.int(1)), tla.enumSet(tla.int(1)))
-    // T = {3} \ 3
-    val set2 = tla.setminus(tla.enumSet(tla.int(3)), tla.enumSet(tla.int(3)))
-    // R = S \cup T = {}
-    val set3 = tla.cup(set1, set2) // the buggy implementation will try in(1, T) and this may return true!
-    val trans = tla.in(tla.int(1), set3)
-    val state = new SymbState(trans, BoolTheory(), arena, new Binding)
-    val rewriter = create()
-    val nextState = rewriter.rewriteUntilDone(state)
-    nextState.ex match {
-      case predEx@NameEx(name) =>
-        solverContext.assertGroundExpr(predEx)
-        assert(!solverContext.sat()) // in the buggy implementation, 1 \ in R
-
-      case _ =>
-        fail("Unexpected rewriting result")
-    }
-  }
-
   test("""SE-SUBSETEQ[1-3]: {1, 2} \subseteq {1, 2, 3} ~~> $B$... (true)""") {
     val left = tla.enumSet(tla.int(1), tla.int(2))
     val right = tla.enumSet(tla.int(1), tla.int(2), tla.int(3))
